removed client side auth

+++ src/lib/compontents/ticketForm.svelte
@@ -4,7 +4,7 @@
 	import { user } from '$lib/stores/userStore';
 
 	import { computers, courses } from '$lib/stores/utilStore';
-	import { supabase } from '$lib/supabaseClient';
+	import { supabaseClient } from '$lib/supabaseClient';
 	import { reporter, ValidationMessage } from '@felte/reporter-svelte';
 	import { validateSchema } from '@felte/validator-zod';
 	import { createForm } from 'felte';
@@ -45,7 +45,8 @@
 		onSubmit: async (values) => {
 			loading = true;
 			try {
-				const { data: ticketData, error: ticketError } = await supabase
+				if (!supabaseClient) throw 'supabase clinet not instantiated';
+				const { data: ticketData, error: ticketError } = await supabaseClient
 					.from('tickets')
 					.insert({
 						title: values.title,
@@ -60,19 +61,21 @@
 				ticketId = ticketData.id;
 				if (ticketError) throw ticketError;
 				if (atComputerLab) {
-					const { error: labTicketError } = await supabase.from('school_tickets').insert({
+					if (!supabaseClient) throw 'supabase clinet not instantiated';
+					const { error: labTicketError } = await supabaseClient.from('school_tickets').insert({
 						ticket_id: ticketData.id,
 						computer_id: computerId
 					});
 					if (labTicketError) throw labTicketError;
 				} else {
-					const { error: homeTicketError } = await supabase.from('home_tickets').insert({
+					if (!supabaseClient) throw 'supabase clinet not instantiated';
+					const { error: homeTicketError } = await supabaseClient.from('home_tickets').insert({
 						ticket_id: ticketData.id,
 						operating_system: operatingSystem
 					});
 					if (homeTicketError) throw homeTicketError;
 				}
-				const { data: messageData, error: messageError } = await supabase
+				const { data: messageData, error: messageError } = await supabaseClient
 					.from('ticket_messages')
 					.insert({
 						ticket_id: ticketData.id,
@@ -85,7 +88,8 @@
 			} catch (e: any) {
 				alert(e.message);
 				if (!ticketId) return;
-				const { data: deleteData, error: deleteError } = await supabase
+				if (!supabaseClient) throw 'supabase clinet not instantiated';
+				const { data: deleteData, error: deleteError } = await supabaseClient
 					.from('tickets')
 					.delete()
 					.select('*')
